editing products, users, orders, myorders pages and sessions and logout

+++ server/adminPages/deleteProduct.php
@@ -1,12 +1,19 @@
 <?php 
 require('../env.php');
 session_start();
-$ID = json_decode(file_get_contents("php://input"), true)['product_id'];
+if (!array_key_exists("admin",$_SESSION))
+{  
+    $notAuthorized = [
+        "notAuthorized"=>true
+    ];
+    echo json_encode($notAuthorized);
 
-$query = "DELETE from products where product_id = $ID";
-$sql = $conn->prepare($query);
-$result = $sql->execute();
-if($result){
-    echo json_encode(["success"=>true]);
+} else {
+    $ID = json_decode(file_get_contents("php://input"), true)['product_id'];
+    $query = "DELETE from products where product_id = $ID";
+    $sql = $conn->prepare($query);
+    $result = $sql->execute();
+    if ($result) {
+        echo json_encode(["success" => true]);
+    }
 }
-
